Support 2 NK2's, alpha bug

+++ core/time.h
@@ -2,7 +2,7 @@
 #define __CORE_TIME_H__
 
 typedef long mbeat_t;
-#define MB2B(x) (float) ((double) (x) / 1000.)
-#define B2MB(x) (long) ((long) ((x) * 1000.))
+#define MB2B(x) ((double) (x) / 1000.)
+#define B2MB(x) ((long) ((x) * 1000.))
 
 #endif
